Make property tests work when xlwings not installed
